--- src/main/resources/customerStorage.json
@@ -1,1 +1,0 @@
-{"personalData":[{"firstName":"Ion","lastName":"Castan","cryptedCard":"4xt+gGsU65DgAmfgqR4TH9MxpE7dN2Q3QCSzEPYbGTE=","clientId":0,"phoneNumber":"0756444890","adress":"Jud. PH,Oras. Bacanesti","email":"youexp@gmail.com","username":"xxDemonSlayerxx","encryptedPassword":"uXiVusagH5J+XR3m4ZeuhA=="},{"firstName":"Ion","lastName":"Castan","cryptedCard":"4xt+gGsU65DgAmfgqR4TH9MxpE7dN2Q3QCSzEPYbGTE=","clientId":1,"phoneNumber":"0756444890","adress":"Jud. PH,Oras. Bacanesti","email":"youexp@gmail.com","username":"xxDemonSlayerxx","encryptedPassword":"uXiVusagH5J+XR3m4ZeuhA=="},{"firstName":"Mihai","lastName":"Corneliu","cryptedCard":"5ANIl4THUOV1ymKjg1zLx9MxpE7dN2Q3QCSzEPYbGTE=","clientId":2,"phoneNumber":"0782443890","adress":"Jud. Timis,Oras. Timisoara","email":"youexp2@gmail.com","username":"Bravo","encryptedPassword":"pUpjy9SALE6cxMf8Pecyew=="},{"firstName":"Maria","lastName":"Castan","cryptedCard":"ENULIHMjistD13dWThCBM/koh75dwr4JtWDiJfIpiS4=","clientId":3,"phoneNumber":"0756420897","adress":"Jud. Timis,Oras. Bacanesti","email":"youexp3@gmail.com","username":"Aistil","encryptedPassword":"YR60buCYXP9VP6Ax3IK7tQ=="}]}
