+++ src/main/resources/staffStorage.json
@@ -1,0 +1,1 @@
+{"personalData":[{"firstName":"Marin","lastName":"Costea","orderList":"[Ljava.lang.Object;@4563e9ab","salary":"2500.0","staffId":0,"email":"interzis@yahoo.com","username":"mar23","encryptedPassword":"F6YUgxfj+a/JKswAOYRqAA=="},{"firstName":"Ioana","lastName":"Gheorghe","orderList":"[Ljava.lang.Object;@11531931","salary":"4000.0","staffId":1,"email":"rockit@gmai.com","username":"nope12","encryptedPassword":"WdtOMGSOOKeaiMv/8CaeaA=="},{"firstName":"Jane","lastName":"Daria","orderList":"[Ljava.lang.Object;@5e025e70","salary":"3000.0","staffId":2,"email":"lol@gmail.com","username":"Dar12","encryptedPassword":"KqFohKZOhKW/sk5ZpZgGxw=="}]}
